refactor: 역할별로 경로에 정리
