+++ .vscode/settings.json
@@ -3,6 +3,6 @@
 	"deno.config": "./deno.jsonc",
 	"[typescript]": {
 		"editor.defaultFormatter": "denoland.vscode-deno"
-	},
-	"typescript.preferGoToSourceDefinition": true
+	}
+	// "typescript.preferGoToSourceDefinition": true
 }
