--- data.ts
@@ -1,4 +1,0 @@
-import { z } from "zod"
-
-export const NumericSchema = z.string().regex(/^\d+$/gm).transform((x) => Number(x))
-export const IntSchema = z.number().int().nonnegative()
