+++ routes/applications.ts
@@ -1,12 +1,13 @@
+import { Hono } from "hono"
 import { zValidator } from "@hono/zod-validator"
 import { PrismaClient } from "@prisma/client"
-import { Hono } from "hono"
-import { ApplicationSchema } from "../prisma/generated/zod"
 
-const applyValidator = zValidator("json", ApplicationSchema.omit({ id: true }))
+import { ApplicationSchema } from "../prisma/generated/zod/index.ts"
+
+const bodyValidator = zValidator("json", ApplicationSchema.omit({ id: true }))
 
 export const applications = (prisma: PrismaClient) =>
-	new Hono().post("/", applyValidator, async (c) => {
+	new Hono().post("/", bodyValidator, async (c) => {
 		const data = c.req.valid("json")
 
 		const result = await prisma.application.create({ data })
