+++ routes/positions/validators/id_validator.ts
@@ -1,5 +1,5 @@
 import { zValidator } from "@hono/zod-validator"
 import { z } from "zod"
-import { NumericSchema } from "../data.js"
 
+export const NumericSchema = z.string().regex(/^\d+$/gm).transform((x) => Number(x))
 export const idValidator = zValidator("param", z.object({ id: NumericSchema }))
