+++ utils/is_main.ts
